Refactor analytics page into helper module
